src/backend/core/ai/voice_transformer.py
import torch
import torchaudio
import numpy as np
from typing import Dict, Any, List
import uuid
import os

class AdvancedVoiceTransformer:
    """
    Advanced AI-powered voice transformation system.
    Supports multiple transformation techniques and styles.
    """
    
    def __init__(self, models_dir='/app/models/voice'):
        """
        Initialize voice transformation models and configurations.
        
        :param models_dir: Directory containing pre-trained models
        """
        self.models_dir = models_dir
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        
        # Load pre-trained models
        self.pitch_model = self._load_pitch_model()
        self.timbre_model = self._load_timbre_model()
        self.gender_model = self._load_gender_model()
    
    def _load_pitch_model(self):
        """
        Load pitch transformation model.
        
        :return: Loaded pitch transformation model
        """
        model_path = os.path.join(self.models_dir, 'pitch_transformer.pth')
        try:
            model = torch.load(model_path, map_location=self.device)
            model.eval()
            return model
        except Exception as e:
            raise ValueError(f"Failed to load pitch model: {e}")
    
    def _load_timbre_model(self):
        """
        Load timbre transformation model.
        
        :return: Loaded timbre transformation model
        """
        model_path = os.path.join(self.models_dir, 'timbre_transformer.pth')
        try:
            model = torch.load(model_path, map_location=self.device)
            model.eval()
            return model
        except Exception as e:
            raise ValueError(f"Failed to load timbre model: {e}")
    
    def _load_gender_model(self):
        """
        Load gender voice transformation model.
        
        :return: Loaded gender transformation model
        """
        model_path = os.path.join(self.models_dir, 'gender_transformer.pth')
        try:
            model = torch.load(model_path, map_location=self.device)
            model.eval()
            return model
        except Exception as e:
            raise ValueError(f"Failed to load gender model: {e}")
    
    def transform_voice(
        self, 
        audio_tensor: torch.Tensor, 
        transformation_params: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Apply advanced voice transformation.
        
        :param audio_tensor: Input audio tensor
        :param transformation_params: Transformation configuration
        :return: Transformed audio and metadata
        """
        try:
            # Pitch transformation
            if transformation_params.get('pitch_shift'):
                audio_tensor = self._apply_pitch_shift(
                    audio_tensor, 
                    transformation_params['pitch_shift']
                )
            
            # Timbre modification
            if transformation_params.get('timbre_style'):
                audio_tensor = self._modify_timbre(
                    audio_tensor, 
                    transformation_params['timbre_style']
                )
            
            # Gender voice transformation
            if transformation_params.get('gender_transform'):
                audio_tensor = self._transform_gender(
                    audio_tensor, 
                    transformation_params['gender_transform']
                )
            
            # Generate unique identifier for transformation
            transform_id = str(uuid.uuid4())
            
            return {
                'transformed_audio': audio_tensor,
                'transform_id': transform_id,
                'metadata': {
                    'device': str(self.device),
                    'transformation_params': transformation_params
                }
            }
        
        except Exception as e:
            raise ValueError(f"Voice transformation failed: {e}")
    
    def _apply_pitch_shift(
        self, 
        audio: torch.Tensor, 
        pitch_shift_semitones: float
    ) -> torch.Tensor:
        """
        Apply pitch shifting to audio.
        
        :param audio: Input audio tensor
        :param pitch_shift_semitones: Semitones to shift pitch
        :return: Pitch-shifted audio tensor
        """
        return self.pitch_model(audio, pitch_shift_semitones)
    
    def _modify_timbre(
        self, 
        audio: torch.Tensor, 
        timbre_style: str
    ) -> torch.Tensor:
        """
        Modify audio timbre based on specified style.
        
        :param audio: Input audio tensor
        :param timbre_style: Desired timbre style
        :return: Timbre-modified audio tensor
        """
        return self.timbre_model(audio, timbre_style)
    
    def _transform_gender(
        self, 
        audio: torch.Tensor, 
        target_gender: str
    ) -> torch.Tensor:
        """
        Transform voice gender characteristics.
        
        :param audio: Input audio tensor
        :param target_gender: Target gender voice characteristics
        :return: Gender-transformed audio tensor
        """
        return self.gender_model(audio, target_gender)
    
    def analyze_voice_characteristics(
        self, 
        audio_tensor: torch.Tensor
    ) -> Dict[str, Any]:
        """
        Analyze detailed voice characteristics.
        
        :param audio_tensor: Input audio tensor
        :return: Detailed voice analysis
        """
        try:
            pitch_analysis = self._analyze_pitch(audio_tensor)
            timbre_analysis = self._analyze_timbre(audio_tensor)
            gender_analysis = self._analyze_gender(audio_tensor)
            
            return {
                'pitch': pitch_analysis,
                'timbre': timbre_analysis,
                'gender_characteristics': gender_analysis
            }
        
        except Exception as e:
            raise ValueError(f"Voice analysis failed: {e}")
    
    def _analyze_pitch(self, audio: torch.Tensor) -> Dict[str, float]:
        """
        Analyze pitch characteristics.
        
        :param audio: Input audio tensor
        :return: Pitch analysis details
        """
        # Placeholder for pitch analysis logic
        return {
            'fundamental_frequency': 220.0,  # Hz
            'pitch_stability': 0.85,
            'pitch_range': (100, 300)  # Hz
        }
    
    def _analyze_timbre(self, audio: torch.Tensor) -> Dict[str, Any]:
        """
        Analyze timbre characteristics.
        
        :param audio: Input audio tensor
        :return: Timbre analysis details
        """
        # Placeholder for timbre analysis logic
        return {
            'brightness': 0.6,
            'harmonicity': 0.75,
            'spectral_centroid': 2500.0,
            'dominant_harmonics': [220, 440, 660]
        }
    
    def _analyze_gender(self, audio: torch.Tensor) -> Dict[str, float]:
        """
        Analyze gender voice characteristics.
        
        :param audio: Input audio tensor
        :return: Gender characteristic probabilities
        """
        # Placeholder for gender analysis logic
        return {
            'masculine_probability': 0.7,
            'feminine_probability': 0.3,
            'neutral_probability': 0.0
        }

# Utility functions for audio processing
def load_audio(file_path: str) -> torch.Tensor:
    """
    Load audio file as a tensor.
    
    :param file_path: Path to audio file
    :return: Audio tensor
    """
    try:
        waveform, sample_rate = torchaudio.load(file_path)
        return waveform
    except Exception as e:
        raise ValueError(f"Failed to load audio file: {e}")

def save_audio(
    audio_tensor: torch.Tensor, 
    file_path: str, 
    sample_rate: int = 44100
) -> None:
    """
    Save audio tensor to file.
    
    :param audio_tensor: Audio tensor to save
    :param file_path: Destination file path
    :param sample_rate: Audio sample rate
    """
    try:
        torchaudio.save(file_path, audio_tensor, sample_rate)
    except Exception as e:
        raise ValueError(f"Failed to save audio file: {e}")
